Refactor directory/file structure
