+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 
 import App from './containers/App'
-import configureStore from './configureStore'
+import configureStore from './store/configureStore'
 
 const store = configureStore()
